refactor(core,auth): move ApiKeyRecord+ApiKeyStore to core::auth
Step 1 of inverting the storage→auth dependency. The trait and record live
in `assistant_core::auth` so that `assistant_storage::SqliteApiKeyStore`
can implement it without taking a path-dep on `assistant_auth`. The
in-memory implementation, key generation, and resolution helpers stay in
`assistant_auth::api_keys` (still re-exporting the moved names for
backwards compatibility).

Refs openspec/changes/resolve-priority-tech-debt/specs/storage-layering/spec.md
